Move `tools` API to the new module.
