+++ tools/tool-base/build.gradle.kts
@@ -24,6 +24,7 @@ group = "io.spine.tools"
 
 dependencies {
     api(project(":base"))
+    api(project(":tools-api"))
     implementation(Deps.gen.javaPoet)
     implementation(Deps.gen.javaxAnnotation)
