+++ tools/tools-api/src/main/java/io/spine/tools/protoc/iface/Is.java
