+++ tools/tools-api/src/main/java/io/spine/tools/protoc/method/GeneratedMethod.java
