+++ tools/tools-api/src/main/java/io/spine/tools/protoc/method/package-info.java
