+++ tools/tools-api/src/main/java/io/spine/tools/protoc/nested/GeneratedNestedClass.java
