+++ tools/tools-api/src/main/java/io/spine/tools/protoc/nested/package-info.java
